+++ CHANGELOG.md
@@ -367,6 +367,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- New class-based resource:
+  - `SqlInstall` - Handles the Microsoft SQL Server setup action `Install`.
 - New public commands:
   - `Disconnect-SqlDscDatabaseEngine` - Disconnects from a SQL Server instance
     that was previously connected to using `Connect-SqlDscDatabaseEngine`.
@@ -381,6 +383,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     for a file.
   - `Assert-Feature` - Throws an exception if a feature is not supported
     for a specific Microsoft SQL Server major version.
+  - `Get-RegistryPropertyValue` - Returns the value of the provided property
+    at the provided registry path.
+  - `ConvertFrom-ServiceStartMode` - Converts the specified start mode to
+    the equivalent normalized startup type.
+  - `Get-InstanceId` - Returns the SQL Server instance id of the specified
+    service type and instance name.
 - SqlServerDsc.Common
   - `Connect-SQL`.
     - Add new parameter `Encrypt`.